+++ packages/sampler/stories/moonstone-stories/VirtualList.js
@@ -1,13 +1,13 @@
-import VirtualList from '@enact/moonstone/VirtualList';
-import ri from '@enact/ui/resolution';
 import Item from '@enact/moonstone/Item';
+import VirtualList from '@enact/moonstone/VirtualList';
 import {VirtualListCore} from '@enact/moonstone/VirtualList/VirtualListBase';
+import ri from '@enact/ui/resolution';
 import React from 'react';
 import {storiesOf, action} from '@kadira/storybook';
 import {boolean, number} from '@kadira/storybook-addon-knobs';
 
-import {mergeComponentMetadata} from '../../src/utils/propTables';
 import nullify from '../../src/utils/nullify.js';
+import {mergeComponentMetadata} from '../../src/utils/propTables';
 
 const Config = mergeComponentMetadata('VirtualList', VirtualListCore, VirtualList);
 
@@ -48,8 +48,8 @@ storiesOf('VirtualList')
 					component={renderItem(itemSize)}
 					data={items}
 					dataSize={number('dataSize', items.length)}
-					itemSize={itemSize}
 					focusableScrollbar={nullify(boolean('focusableScrollbar', false))}
+					itemSize={itemSize}
 					onScrollStart={action('onScrollStart')}
 					onScrollStop={action('onScrollStop')}
 					spacing={ri.scale(number('spacing', 0))}