+++ packages/moonstone/CHANGELOG.md
@@ -6,17 +6,33 @@ The following is a curated list of changes in the Enact moonstone module, newest
 
 ### Deprecated
 
+- `moonstone/Input` prop `noDecorator` is being replaced by `autoFocus` in 2.0.0.
+
 ### Added
 
-`moonstone/styles/text.less` mixin `.locale-japanese-line-break()` to apply the correct  Japanese language line-break rules for the following multi-line components: `moonstone/BodyText`, `moonstone/Dialog`, `moonstone/Notification`, `moonstone/Popup`, and `moonstone/Tooltip`
+- `moonstone/styles/text.less` mixin `.locale-japanese-line-break()` to apply the correct  Japanese language line-break rules for the following multi-line components: `moonstone/BodyText`, `moonstone/Dialog`, `moonstone/Notification`, `moonstone/Popup`, and `moonstone/Tooltip`
+- `moonstone/ContextualPopupDecorator` property `popupProps` to attach props to popup component
 
 ### Changed
 
+- `moonstone/Panels/Header` to expose its `marqueeOn` prop
+- `moonstone/VideoPlayer` to automatically adjust the width of the allocated space for the side components so the media controls have more space to appear on smaller screens
+- `moonstone/VideoPlayer` properties `autoCloseTimeout` and `titleHideDelay` default value to `5000`
+
 ### Fixed
 
-- `moonstone/Dialog' to properly focus via pointer on child components
+- `moonstone/Dialog` to properly focus via pointer on child components
+- `moonstone` component hold delay time
 
 ### Removed
+
+## [1.3.1] - 2017-06-14
+
+### Fixed
+
+- `moonstone/Picker` support for large text
+- `moonstone/Scroller` support for focusing paging controls with the pointer
+- `moonstone` CSS rules for unskinned spottable components
 
 ## [1.3.0] - 2017-06-12